fix the location of pill icon
